1. 更改ECPM类型

+++ adsdk/src/main/java/com/inner/adsdk/adloader/base/AbstractSdkLoader.java
@@ -599,7 +599,7 @@ public abstract class AbstractSdkLoader implements ISdkLoader, Handler.Callback 
     }
 
     @Override
-    public int getEcpm() {
+    public double getEcpm() {
         if (mPidConfig != null) {
             return mPidConfig.getEcpm();
         }
